user 보유 캐시 int -> float 변경

+++ src/main/java/bit/bitgroundspring/util/InitialCashUtil.java
@@ -11,7 +11,7 @@ public class InitialCashUtil {
     @Value("${scheduler.setting}")
     private String schedulerSetting;
     
-    public int getInitialCash() {
+    public float getInitialCash() {
         LocalDateTime now = LocalDateTime.now();
         int day = now.getDayOfMonth();
         int hour = now.getHour();
